+++ src/tr1/game/lara/cheat.c
@@ -376,7 +376,12 @@ bool Lara_Cheat_KillEnemy(const int16_t item_num)
     Item_Explode(item_num, -1, 0);
     Sound_Effect(SFX_EXPLOSION_CHEAT, &item->pos, SPM_NORMAL);
     Item_Kill(item_num);
-    LOT_DisableBaddieAI(item_num);
+    if (item->object_id != O_MUMMY) {
+        // The Qualopec mummy uses only one short as its creature data; having
+        // LOT disable this can lead to memory corruption as it expects a
+        // regular CREATURE.
+        LOT_DisableBaddieAI(item_num);
+    }
     item->flags |= IF_ONE_SHOT;
     Carrier_TestItemDrops(item_num);
     return true;